Rename BaseAppConfig to be snake_case
